Implemented Stack -- Fixed & Tested
Finished the stack implementation.
Look at stack_test.c as an example for how to
initialize, destroy, or use the functions for this container.

Note: B/C I used void pointers, you could technically really screw up
your program by passing in memory that shouldn't be passed. Be careful
when using  generic functions like these.
